backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-community
huggingface_hub
python-dotenv
pydantic